server/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
mysql-connector-python==8.1.0
redis==5.0.1
python-dotenv==1.0.0
eventlet==0.33.3
gunicorn==21.2.0
watchdog==3.0.0
requests==2.31.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.26
openai>=1.0.0
